Move parse_md_file into config.py; delete frontmatter.py
